requirements.txt
selenium>=3.141.0
arabic-reshaper==2.1.0
python-bidi==0.4.2
Pillow==6.2.1